refactor: extract ArtistCard leaf component (Phase 5.2 follow-up)
Mirrors the AlbumCard/TrackList memoization pattern: pulls the
artist grid-cell markup out of LibraryArtistsSection into its own
React.memo'd component with a stable onClick via useCallback.
